+++ TestHost/IPC/IPCMessage.h
@@ -80,7 +80,7 @@ public:
 
     // append keyed argument to the message
     template <typename ArgT>
-    void append (size_t argKey, ArgT argValue)
+    void append (int32_t argKey, ArgT argValue)
     {
         _appendArg (argKey, argValue);
     }
@@ -88,14 +88,14 @@ public:
     // getters for receiving code: extract arguments, specifying the desired return type as template argument
     // the optional variant returns the default value of the argument type if key wasn't not found
     template <typename ArgT>
-    ArgT getArgValue (size_t argKey) const
+    ArgT getArgValue (int32_t argKey) const
     {
         ArgT result;
         _readArg (argKey, result);
         return result;
     }
     template <typename ArgT>
-    std::pair<ArgT, bool> getOptionalArgValue (size_t argKey) const
+    std::pair<ArgT, bool> getOptionalArgValue (int32_t argKey) const
     {
         ArgT result;
         bool didFindKey { false };
@@ -109,38 +109,38 @@ private:
     void _setup (CFDictionaryRef dictionary, bool isWritable);
 
     // wrap key value into CFString (no reference count transferred to caller)
-    static CFStringRef _getEncodedKey (size_t argKey);
+    static CFStringRef _getEncodedKey (int32_t argKey);
 #else
-    static const char* _getEncodedKey (size_t argKey);
+    static const char* _getEncodedKey (int32_t argKey);
 #endif
 
     // helpers for construction for sending code
-    void _appendArg (size_t argKey, int32_t argValue);
-    void _appendArg (size_t argKey, int64_t argValue);
-    void _appendArg (size_t argKey, size_t argValue);
-    void _appendArg (size_t argKey, float argValue);
-    void _appendArg (size_t argKey, double argValue);
-    void _appendArg (size_t argKey, const char* argValue);
-    void _appendArg (size_t argKey, const std::vector<uint8_t>& argValue);
-    void _appendArg (size_t argKey, const IPCMessage& argValue);
+    void _appendArg (int32_t argKey, int32_t argValue);
+    void _appendArg (int32_t argKey, int64_t argValue);
+    void _appendArg (int32_t argKey, size_t argValue);
+    void _appendArg (int32_t argKey, float argValue);
+    void _appendArg (int32_t argKey, double argValue);
+    void _appendArg (int32_t argKey, const char* argValue);
+    void _appendArg (int32_t argKey, const std::vector<uint8_t>& argValue);
+    void _appendArg (int32_t argKey, const IPCMessage& argValue);
 
 #if IPC_MESSAGE_USE_CFDICTIONARY
-    void _appendEncodedArg (size_t argKey, __attribute__((cf_consumed)) CFTypeRef argObject);
+    void _appendEncodedArg (int32_t argKey, __attribute__((cf_consumed)) CFTypeRef argObject);
 #else
     void _makeWritableIfNeeded ();
-    pugi::xml_attribute _appendAttribute (size_t argKey);
+    pugi::xml_attribute _appendAttribute (int32_t argKey);
 #endif
 
     // helpers for getters for receiving code
     // since we cannot overload by return type, we are returning via reference argument
-    void _readArg (size_t argKey, int32_t& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, int64_t& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, size_t& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, float& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, double& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, const char*& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, std::vector<uint8_t>& argValue, bool* didFindKey = nullptr) const;
-    void _readArg (size_t argKey, IPCMessage& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, int32_t& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, int64_t& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, size_t& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, float& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, double& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, const char*& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, std::vector<uint8_t>& argValue, bool* didFindKey = nullptr) const;
+    void _readArg (int32_t argKey, IPCMessage& argValue, bool* didFindKey = nullptr) const;
 
     static bool _checkKeyFound (bool found, bool* didFindKey)
     {
